src/main/java/com/weather/controller/WeatherController.java
package com.weather.controller;

import com.weather.dto.WeatherResponse;
import com.weather.service.WeatherService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;

import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.time.LocalDate;

@RestController
@RequestMapping("/weather")
@RequiredArgsConstructor
@Slf4j
public class WeatherController {
    private final WeatherService weatherService;

    @GetMapping
    public ResponseEntity<WeatherResponse> getWeatherInfo(
        @RequestParam String pincode,
        @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate forDate) 
    {
    	log.debug("Attempting to fetch weather for pincode: {} on date: {}", pincode, forDate);
        return ResponseEntity.ok(weatherService.getWeatherForPincodeAndDate(pincode, forDate));
    }
}